requirements.txt
asgiref==3.5.2
beautifulsoup4==4.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==4.0.2
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
mysqlclient==2.1.0
oauthlib==3.2.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-decouple==3.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
types-cryptography==3.3.23
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.0.0
zipp==3.8.1
